faucet-backend/requirements.txt
fastapi
web3
uvicorn
pyjwt
